# docs/swagger.yaml
basePath: /
definitions:
...
  internal_adapters_primary_api_product.addProductReq:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock_qty:
        type: integer
    type: object
  internal_adapters_primary_api_product.addProductResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      sku:
        type: string
    type: object
  internal_adapters_primary_api_product.editProductReq:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock_qty:
        type: integer
    type: object
  internal_adapters_primary_api_product.editProductResp:
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock_qty:
        type: integer
    type: object
  internal_adapters_primary_api_product.getProductResp:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock_qty:
        type: integer
    type: object
  internal_adapters_primary_api_product.listProductsResp:
    properties:
      products:
        items:
          $ref: '#/definitions/internal_adapters_primary_api_product.productInfoResp'
        type: array
    type: object
  internal_adapters_primary_api_product.productInfoResp:
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock_qty:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    A RESTful e-commerce API built with Go using hexagonal architecture.
    Features: User authentication (JWT), Product management, Order processing, Address management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
...
  /admin/products:
    post:
      consumes:
      - application/json
      description: Create a new product (admin only)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_product.addProductReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_product.addProductResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add a new product (Admin)
      tags:
      - Admin
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a product (Admin)
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update product details (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_product.editProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_product.editProductResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Edit a product (Admin)
      tags:
      - Admin
...
  /products:
    get:
      consumes:
      - application/json
      description: Get a list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_product.listProductsResp'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List all products
      tags:
      - Products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Get product details by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_product.getProductResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Get a product
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"